server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Brooke Chalmers <breq@breq.dev>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
Flask-SQLAlchemy = "^2.5.1"
python-dotenv = "^0.18.0"
PyJWT = "^2.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
